factor out COSE-key helpers common between both backends
