+++ passkey-crypto/src/aws_lc_rs/mod.rs
@@ -7,7 +7,13 @@ use coset::{
 use std::ops::RangeInclusive;
 
 use crate::{
-    CoseKeyConversionError, CryptoBackend, PublicKeyT, SecretKeyT, hash::Sha256Backend,
+    CoseKeyConversionError, CryptoBackend, PublicKeyT, SecretKeyT,
+    cose::{
+        ED25519_KEY_LEN, P256_UNCOMPRESSED_LEN, extract_okp_d, extract_okp_x,
+        extract_p256_d, extract_p256_xy, find_ec2_crv, find_okp_crv, find_okp_x, p256_uncompressed,
+        split_p256_uncompressed,
+    },
+    hash::Sha256Backend,
     rng::RngBackend,
 };
 use aws_lc_rs::{
@@ -20,14 +26,6 @@ use aws_lc_rs::{
         Ed25519KeyPair, KeyPair, ParsedPublicKey, UnparsedPublicKey,
     },
 };
-use zeroize::Zeroizing;
-
-/// Length of an uncompressed SEC1 encoded P-256 public key (0x04 || X || Y).
-const P256_UNCOMPRESSED_LEN: usize = 65;
-/// Length of a P-256 field element (X or Y coordinate, or the D scalar).
-const P256_FIELD_LEN: usize = 32;
-/// Length of an Ed25519 seed / public key.
-const ED25519_KEY_LEN: usize = 32;
 
 /// Secret key backed by aws-lc-rs.
 pub struct AwsLcRsSecretKey(AwsLcRsSecretKeyInner);
@@ -364,159 +362,3 @@ impl CryptoBackend for AwsLcRsBackend {
     }
 }
 
-fn extract_p256_xy(
-    cose_key: &CoseKey,
-) -> Result<([u8; P256_FIELD_LEN], [u8; P256_FIELD_LEN]), CoseKeyConversionError> {
-    let (mut x, mut y) = (None, None);
-    for (key, value) in &cose_key.params {
-        if let coset::Label::Int(i) = key {
-            let key = iana::Ec2KeyParameter::from_i64(*i)
-                .ok_or(CoseKeyConversionError::InvalidCredential)?;
-            match key {
-                iana::Ec2KeyParameter::X => {
-                    if value.as_bytes().and_then(|v| x.replace(v)).is_some() {
-                        log::warn!("Cose key has multiple entries for X coordinate");
-                    }
-                }
-                iana::Ec2KeyParameter::Y => {
-                    if value.as_bytes().and_then(|v| y.replace(v)).is_some() {
-                        log::warn!("Cose key has multiple entries for Y coordinate");
-                    }
-                }
-                _ => (),
-            }
-        }
-    }
-    let (Some(x), Some(y)) = (x, y) else {
-        return Err(CoseKeyConversionError::InvalidCredential);
-    };
-    let x: [u8; P256_FIELD_LEN] = x
-        .as_slice()
-        .try_into()
-        .map_err(|_| CoseKeyConversionError::InvalidCredential)?;
-    let y: [u8; P256_FIELD_LEN] = y
-        .as_slice()
-        .try_into()
-        .map_err(|_| CoseKeyConversionError::InvalidCredential)?;
-    Ok((x, y))
-}
-
-fn extract_p256_d(
-    cose_key: &CoseKey,
-) -> Result<Zeroizing<[u8; P256_FIELD_LEN]>, CoseKeyConversionError> {
-    let bytes = cose_key
-        .params
-        .iter()
-        .find_map(|(k, v)| {
-            if let coset::Label::Int(i) = k {
-                iana::Ec2KeyParameter::from_i64(*i)
-                    .filter(|p| p == &iana::Ec2KeyParameter::D)
-                    .and_then(|_| v.as_bytes())
-            } else {
-                None
-            }
-        })
-        .ok_or(CoseKeyConversionError::InvalidCredential)?;
-    if bytes.len() != P256_FIELD_LEN {
-        return Err(CoseKeyConversionError::InvalidCredential);
-    }
-    let mut d: Zeroizing<[u8; P256_FIELD_LEN]> = Zeroizing::new([0u8; P256_FIELD_LEN]);
-    d.copy_from_slice(bytes);
-    Ok(d)
-}
-
-fn extract_okp_x(cose_key: &CoseKey) -> Result<[u8; ED25519_KEY_LEN], CoseKeyConversionError> {
-    find_okp_x(cose_key)?.ok_or(CoseKeyConversionError::InvalidCredential)
-}
-
-fn find_okp_x(cose_key: &CoseKey) -> Result<Option<[u8; ED25519_KEY_LEN]>, CoseKeyConversionError> {
-    let mut x = None;
-    for (key, value) in &cose_key.params {
-        if let coset::Label::Int(i) = key {
-            let key = iana::OkpKeyParameter::from_i64(*i)
-                .ok_or(CoseKeyConversionError::InvalidCredential)?;
-            if key == iana::OkpKeyParameter::X
-                && value.as_bytes().and_then(|v| x.replace(v)).is_some()
-            {
-                log::warn!("Cose key has multiple entries for X coordinate");
-            }
-        }
-    }
-    match x {
-        Some(bytes) => bytes
-            .as_slice()
-            .try_into()
-            .map(Some)
-            .map_err(|_| CoseKeyConversionError::InvalidCredential),
-        None => Ok(None),
-    }
-}
-
-fn find_ec2_crv(cose_key: &CoseKey) -> Result<Option<i64>, CoseKeyConversionError> {
-    find_crv(cose_key, iana::Ec2KeyParameter::Crv.to_i64())
-}
-
-fn find_okp_crv(cose_key: &CoseKey) -> Result<Option<i64>, CoseKeyConversionError> {
-    find_crv(cose_key, iana::OkpKeyParameter::Crv.to_i64())
-}
-
-fn find_crv(cose_key: &CoseKey, crv_label: i64) -> Result<Option<i64>, CoseKeyConversionError> {
-    let Some(value) = cose_key.params.iter().find_map(|(k, v)| match k {
-        coset::Label::Int(i) if *i == crv_label => Some(v),
-        _ => None,
-    }) else {
-        return Ok(None);
-    };
-    let crv: i64 = value
-        .as_integer()
-        .ok_or(CoseKeyConversionError::InvalidCredential)?
-        .try_into()
-        .map_err(|_| CoseKeyConversionError::InvalidCredential)?;
-    Ok(Some(crv))
-}
-
-fn extract_okp_d(
-    cose_key: &CoseKey,
-) -> Result<Zeroizing<[u8; ED25519_KEY_LEN]>, CoseKeyConversionError> {
-    let bytes = cose_key
-        .params
-        .iter()
-        .find_map(|(k, v)| {
-            let coset::Label::Int(i) = k else {
-                return None;
-            };
-            iana::OkpKeyParameter::from_i64(*i)
-                .filter(|p| *p == iana::OkpKeyParameter::D)
-                .and_then(|_| v.as_bytes())
-        })
-        .ok_or(CoseKeyConversionError::InvalidCredential)?;
-    if bytes.len() != ED25519_KEY_LEN {
-        return Err(CoseKeyConversionError::InvalidCredential);
-    }
-    let mut seed: Zeroizing<[u8; ED25519_KEY_LEN]> = Zeroizing::new([0u8; ED25519_KEY_LEN]);
-    seed.copy_from_slice(bytes);
-    Ok(seed)
-}
-
-fn p256_uncompressed(
-    x: &[u8; P256_FIELD_LEN],
-    y: &[u8; P256_FIELD_LEN],
-) -> [u8; P256_UNCOMPRESSED_LEN] {
-    let mut out = [0u8; P256_UNCOMPRESSED_LEN];
-    out[0] = 0x04;
-    out[1..1 + P256_FIELD_LEN].copy_from_slice(x);
-    out[1 + P256_FIELD_LEN..].copy_from_slice(y);
-    out
-}
-
-fn split_p256_uncompressed(
-    bytes: &[u8; P256_UNCOMPRESSED_LEN],
-) -> (&[u8; P256_FIELD_LEN], &[u8; P256_FIELD_LEN]) {
-    let x = (&bytes[1..1 + P256_FIELD_LEN])
-        .try_into()
-        .expect("slice length matches array");
-    let y = (&bytes[1 + P256_FIELD_LEN..])
-        .try_into()
-        .expect("slice length matches array");
-    (x, y)
-}
